cleanup : clean ecs system
